added ability to set offset using data in CoefficientSet; updated examples, tests, and usage
