+++ examples/ex_01_quickstart.py
@@ -14,7 +14,6 @@ max_coefficient = 5                                         # value of largest/s
 max_L0_value = 5                                            # maximum model size (set as float(inf))
 max_offset = 50                                             # maximum value of offset parameter (optional)
 c0_value = 1e-6                                             # L0-penalty parameter such that c0_value > 0; larger values -> sparser models; we set to a small value (1e-6) so that we get a model with max_L0_value terms
-w_pos = 1.00                                                # relative weight on examples with y = +1; w_neg = 1.00 (optional)
 
 
 # load data from disk
@@ -22,10 +21,7 @@ data = riskslim.load_data_from_csv(dataset_csv_file = data_csv_file, sample_weig
 
 # create coefficient set and set the value of the offset parameter
 coef_set = riskslim.CoefficientSet(variable_names = data['variable_names'], lb = -max_coefficient, ub = max_coefficient, sign = 0)
-conservative_offset = riskslim.get_conservative_offset(data, coef_set, max_L0_value)
-max_offset = min(max_offset, conservative_offset)
-coef_set['(Intercept)'].ub = max_offset
-coef_set['(Intercept)'].lb = -max_offset
+coef_set.update_intercept_bounds(X = data['X'], y = data['Y'], max_offset = max_offset)
 
 constraints = {
     'L0_min': 0,
@@ -37,7 +33,6 @@ constraints = {
 settings = {
     # Problem Parameters
     'c0_value': c0_value,
-    'w_pos': w_pos,
     #
     # LCPA Settings
     'max_runtime': 30.0,                               # max runtime for LCPA
