+++ riskslim/__init__.py
@@ -1,4 +1,4 @@
 from .coefficient_set import CoefficientSet
 from .lattice_cpa import run_lattice_cpa, setup_lattice_cpa, finish_lattice_cpa
 from .helper_functions import load_data_from_csv, print_model
-from .setup_functions import get_conservative_offset
+from riskslim.coefficient_set import get_conservative_offset
