--- riskslim/tests/test_coefficient_set.py
@@ -1,3 +1,0 @@
-#noinspection
-
-#todo
